ecosystem: break sidebar + archive + ecosystem directories into components
