+++ pages/ecosystem/index.js
@@ -6,7 +6,6 @@ import {
   SingleColumn,
   Section,
   IntraNav,
-  Markdown,
   getAllPosts,
   generateDisplayDate,
   formatDate,
@@ -14,6 +13,7 @@ import {
 import Header from "../../components/Header";
 import Footer from "../../components/Footer";
 import Sidebar from "../../components/Sidebar";
+import SpotlightArchive from "../../components/ecosystem/SpotlightArchive";
 import Post from "../../components/ecosystem/Post";
 import classnames from "classnames";
 import Link from "next/link";
@@ -37,9 +37,14 @@ export default function Ecosystem({
 
   const { type } = router.query;
 
+  // Route ?type=archive to below component;
+  // otherwise, show directory of ecosystem type
+  // If no param, we show the spotlight for the month
   if (type === "archive") {
     return <SpotlightArchive posts={posts} search={search} />;
   }
+
+  const spotlight = type === undefined;
 
   let title;
   if (type) {
@@ -92,21 +97,19 @@ export default function Ecosystem({
             </Sidebar>
             <div
               className={classnames("w-full", {
-                "max-w-prose": type === undefined,
+                "max-w-prose": spotlight,
               })}
             >
-              {type === undefined && (
-                <h3>{`Ecosystem Spotlight - ${post.title}`}</h3>
-              )}
-              {type === undefined && (
+              {/* Default layout */}
+              {spotlight && (
                 <>
+                  <h3>{`Ecosystem Spotlight - ${post.title}`}</h3>
                   <Post post={post} />
                   <Link href="/ecosystem/?type=archive" passHref>
                     <a className="button-lg bg-wall-600 text-white w-fit">
                       Spotlight Archive
                     </a>
                   </Link>
-
                   <div className="pt-36">
                     <h3 className="pb-2">Urbit Monthly</h3>
                     <p className="pb-8">
@@ -117,124 +120,36 @@ export default function Ecosystem({
                   </div>
                 </>
               )}
-
+              {/* Ecosystem type layouts */}
               <div
                 className={classnames("grid gap-12 w-full", {
                   "grid-cols-2 md:grid-cols-3": type == "organizations" || type == "marketplaces",
                   "grid-cols-1": type === "podcasts" || type === "articles",
                   "grid-cols-1 md:grid-cols-1 lg:grid-cols-1 xl:grid-cols-2":
                     type === "applications" || type === "groups",
-                  hidden: type === undefined,
+                  hidden: spotlight,
                 })}
               >
                 {type === "organizations" &&
-                  organizations.map((org) => (
-                    <Link href={`/organizations/${org.slug}`}>
-                      <div className="flex flex-col space-y-4  items-center cursor-pointer hover:opacity-90">
-                        <img className="w-36 rounded-xl" src={org.image} />
-                        <p className="text-center font-bold">{org.title}</p>
-                      </div>
-                    </Link>
-                  ))}
+                  <Organizations dir={organizations} />
+                }
 
                 {type === "applications" &&
-                  applications.map((app) => (
-                    <Link href={`/applications/${app.ship}/${app.slug}`}>
-                      <div className="flex space-x-4 items-center cursor-pointer hover:opacity-90">
-                        <div
-                          className="h-36 w-36 rounded-xl items-center justify-center shrink-0 overflow-hidden"
-                          style={{
-                            backgroundColor: app?.bgColor || "rgba(0,0,0,0)",
-                          }}
-                        >
-                          {app?.image && (
-                            <img
-                              className="items-center h-36 w-36"
-                              src={app.image}
-                            />
-                          )}
-                        </div>
-                        <div className="">
-                          <p className="text-left font-bold">{app.title}</p>
-                          <p className="text-left text-wall-400 font-light">
-                            {app.description}
-                          </p>
-                        </div>
-                      </div>
-                    </Link>
-                  ))}
+                  <Applications dir={applications} />
+                }
                 {type === "groups" &&
-                  groups.map((group) => (
-                    <Link href={`/groups/${group.ship}/${group.slug}`}>
-                      <div className="flex space-x-4 items-center cursor-pointer hover:opacity-90">
-                        <div
-                          className="h-36 w-36 rounded-xl items-center justify-center shrink-0 overflow-hidden"
-                          style={{
-                            backgroundColor: group?.tile?.startsWith("#")
-                              ? group?.tile
-                              : "rgba(0,0,0,0)",
-                          }}
-                        >
-                          {!group?.tile?.startsWith("#") && (
-                            <img
-                              className="items-center h-36 w-36"
-                              src={group.tile}
-                            />
-                          )}
-                        </div>
-                        <div className="">
-                          <p className="text-left font-bold">{group.title}</p>
-                          <p className="text-left text-wall-400 font-light">
-                            {group.description}
-                          </p>
-                        </div>
-                      </div>
-                    </Link>
-                  ))}
+                  <Groups dir={groups} />
+                }
                 {type === "marketplaces" &&
-                  marketplaces.map((market) => (
-                    <Link href={`/marketplaces/${market.slug}`}>
-                      <div className="flex flex-col space-y-4 justify-center items-center cursor-pointer hover:opacity-90">
-                        <img className="w-36 rounded-xl" src={market.image} />
-                        <p className="text-center font-bold">{market.title}</p>
-                      </div>
-                    </Link>
-                  ))}
+                  <Marketplaces dir={marketplaces} />
+                }
                 {type === "podcasts" &&
-                  podcasts.map((pod) => (
-                    <Link href={`/podcasts/${pod.slug}`}>
-                      <div className="flex cursor-pointer space-x-4 items-center hover:opacity-90">
-                        <img className="w-28" src={pod.image} />
-                        <div className="flex flex-col space-y-2 min-w-0">
-                          <p className="font-bold">{pod.podcast}</p>
-                          <p className="min-w-0 min-h-0 leading-5 ">
-                            {pod.title}
-                          </p>
-                          <p className="text-wall-400">
-                            {formatDate(generateDisplayDate(pod.date))}
-                          </p>
-                        </div>
-                      </div>
-                    </Link>
-                  ))}
+                  <Podcasts dir={podcasts} />
+                }
 
                 {type === "articles" &&
-                  articles.map((article) => (
-                    <Link href={`/articles/${article.slug}`}>
-                      <div className="flex cursor-pointer space-x-4 items-center hover:opacity-90">
-                        <img className="w-28" src={article.image} />
-                        <div className="flex flex-col space-y-2 min-w-0">
-                          <p className="font-bold">{article.publication}</p>
-                          <p className="min-w-0 min-h-0 leading-5 ">
-                            {article.title}
-                          </p>
-                          <p className="text-wall-400">
-                            {formatDate(generateDisplayDate(article.date))}
-                          </p>
-                        </div>
-                      </div>
-                    </Link>
-                  ))}
+                  <Articles dir={articles} />
+                }
               </div>
             </div>
           </div>
@@ -245,149 +160,132 @@ export default function Ecosystem({
   );
 }
 
-function ActiveLink({ children, href, className, currentPath }) {
-  // special casing this behaviour which overloads one nav
-  const archive = Boolean(
-    currentPath === "/ecosystem?type=archive" && href === "/ecosystem"
-  );
-  const activeClassName = classnames({
-    "text-wall-800": currentPath === href || archive,
-    "text-wall-400": currentPath !== href && !archive,
-  });
-
-  return (
-    <Link href={href} passHref>
-      <a className={`${className} ${activeClassName}`}>{children}</a>
+const Organizations = ({ dir }) => {
+  return dir.map((org) => (
+    <Link href={`/organizations/${org.slug}`}>
+      <div className="flex flex-col space-y-4  items-center cursor-pointer hover:opacity-90">
+        <img className="w-36 rounded-xl" src={org.image} />
+        <p className="text-center font-bold">{org.title}</p>
+      </div>
     </Link>
-  );
+  ))
 }
 
-function EcosystemSidebar() {
-  const router = useRouter();
-  let currentPath = router.asPath;
-  return (
-    <div className="flex flex-col space-y-4 pr-24">
-      <ActiveLink
-        currentPath={currentPath}
-        className="type-ui"
-        href="/ecosystem"
-      >
-        Spotlight
-      </ActiveLink>
-      <ActiveLink
-        currentPath={currentPath}
-        className="type-ui"
-        href="/ecosystem?type=organizations"
-      >
-        Organizations
-      </ActiveLink>
-      <ActiveLink
-        currentPath={currentPath}
-        className="type-ui"
-        href="/ecosystem?type=applications"
-      >
-        Applications
-      </ActiveLink>
-      <ActiveLink
-        currentPath={currentPath}
-        className="type-ui"
-        href="/ecosystem?type=groups"
-      >
-        Groups
-      </ActiveLink>
-      <ActiveLink
-        currentPath={currentPath}
-        className="type-ui"
-        href="/ecosystem?type=groups"
-      >
-        Marketplaces
-      </ActiveLink>
-      <ActiveLink
-        currentPath={currentPath}
-        className="mr-5 type-ui"
-        href="/ecosystem?type=podcasts"
-      >
-        Podcasts
-      </ActiveLink>
-
-      <ActiveLink
-        currentPath={currentPath}
-        className="mr-5 type-ui"
-        href="/ecosystem?type=articles"
-      >
-        Articles
-      </ActiveLink>
-
-
-    </div>
-  );
+const Applications = ({ dir }) => {
+  return dir.map((app) => (
+    <Link href={`/applications/${app.ship}/${app.slug}`}>
+      <div className="flex space-x-4 items-center cursor-pointer hover:opacity-90">
+        <div
+          className="h-36 w-36 rounded-xl items-center justify-center shrink-0 overflow-hidden"
+          style={{
+            backgroundColor: app?.bgColor || "rgba(0,0,0,0)",
+          }}
+        >
+          {app?.image && (
+            <img
+              className="items-center h-36 w-36"
+              src={app.image}
+            />
+          )}
+        </div>
+        <div className="">
+          <p className="text-left font-bold">{app.title}</p>
+          <p className="text-left text-wall-400 font-light">
+            {app.description}
+          </p>
+        </div>
+      </div>
+    </Link>
+  ))
 }
 
-function SpotlightArchive({ posts, search }) {
-  const post = {
-    title: "Spotlight Archive",
-    description: "The archive of ecosystem spotlights.",
-  };
-  return (
-    <Container>
-      <Head>
-        <title>Archive • Spotlight • Urbit</title>
-        {Meta(post)}
-      </Head>
-      <IntraNav ourSite="https://urbit.org" search={search} />
-      <SingleColumn>
-        <Header search={search} />
-        <Section>
-          <div className="measure">
-            <h1>Ecosystem</h1>
-          </div>
-        </Section>
-        <Section className="flex">
-          <Sidebar>
-            <EcosystemSidebar />
-          </Sidebar>
-          <div>
-            <h3 className="mb-12">Spotlight Archive</h3>
-            {posts.map((post) => {
-              return (
-                <div key={post.slug} className="mb-24 cursor-pointer">
-                  <Link href={`/ecosystem/spotlight/${post.slug}`}>
-                    <div>
-                      <h3 className="text-green-400">{post.title}</h3>
-                      <div className="type-ui mt-2">
-                        {["featured-1", "featured-2", "featured-3"].map(
-                          (feat, i) => {
-                            if (post?.[feat]) {
-                              return (
-                                <>
-                                  {i ? ", " : ""}
-                                  <span>{post[feat].title}</span>
-                                </>
-                              );
-                            }
-                          }
-                        )}
-                      </div>
-                    </div>
-                  </Link>
-                </div>
-              );
-            })}
-          </div>
-        </Section>
-      </SingleColumn>
-      <Footer />
-    </Container>
-  );
+const Groups = ({ dir }) => {
+  return dir.map((group) => (
+    <Link href={`/groups/${group.ship}/${group.slug}`}>
+      <div className="flex space-x-4 items-center cursor-pointer hover:opacity-90">
+        <div
+          className="h-36 w-36 rounded-xl items-center justify-center shrink-0 overflow-hidden"
+          style={{
+            backgroundColor: group?.tile?.startsWith("#")
+              ? group?.tile
+              : "rgba(0,0,0,0)",
+          }}
+        >
+          {!group?.tile?.startsWith("#") && (
+            <img
+              className="items-center h-36 w-36"
+              src={group.tile}
+            />
+          )}
+        </div>
+        <div className="">
+          <p className="text-left font-bold">{group.title}</p>
+          <p className="text-left text-wall-400 font-light">
+            {group.description}
+          </p>
+        </div>
+      </div>
+    </Link>
+  ))
 }
 
-export async function getStaticProps({ }) {
+const Marketplaces = ({ dir }) => {
+  return dir.map((market) => (
+    <Link href={`/marketplaces/${market.slug}`}>
+      <div className="flex flex-col space-y-4 justify-center items-center cursor-pointer hover:opacity-90">
+        <img className="w-36 rounded-xl" src={market.image} />
+        <p className="text-center font-bold">{market.title}</p>
+      </div>
+    </Link>
+  ))
+}
+
+const Podcasts = ({ dir }) => {
+  return dir.map((pod) => (
+    <Link href={`/podcasts/${pod.slug}`}>
+      <div className="flex cursor-pointer space-x-4 items-center hover:opacity-90">
+        <img className="w-28" src={pod.image} />
+        <div className="flex flex-col space-y-2 min-w-0">
+          <p className="font-bold">{pod.podcast}</p>
+          <p className="min-w-0 min-h-0 leading-5 ">
+            {pod.title}
+          </p>
+          <p className="text-wall-400">
+            {formatDate(generateDisplayDate(pod.date))}
+          </p>
+        </div>
+      </div>
+    </Link>
+  ))
+}
+
+const Articles = ({ dir }) => {
+  return dir.map((article) => (
+    <Link href={`/articles/${article.slug}`}>
+      <div className="flex cursor-pointer space-x-4 items-center hover:opacity-90">
+        <img className="w-28" src={article.image} />
+        <div className="flex flex-col space-y-2 min-w-0">
+          <p className="font-bold">{article.publication}</p>
+          <p className="min-w-0 min-h-0 leading-5 ">
+            {article.title}
+          </p>
+          <p className="text-wall-400">
+            {formatDate(generateDisplayDate(article.date))}
+          </p>
+        </div>
+      </div>
+    </Link>
+  ))
+}
+
+
+export async function getStaticProps() {
   const posts = getAllPosts(
     ["title", "date", "slug", "featured-1", "featured-2", "featured-3"],
     "ecosystem/spotlight",
     "date"
   );
-  let post = matchEcosystemPost(getAllPosts(
+  const post = matchEcosystemPost(getAllPosts(
     ["title", "date", "featured-1", "featured-2", "featured-3"],
     "ecosystem/spotlight",
     "date"
